Login manager 100Hz: script cleanup
Rewrote comments so the script doesn't emntion login managers.
Since it can be used in any situation, not just on login.

Change variable name resolution_preferred to resolution_100hz.
I think it makes the code easier to understand.

xrandr is now only queried once. The result is cached and reused.
This makes the script easier to debug.